Update README.md
These measures depict a set of data by identifying a central element.
i.	Mean or the arithmetic mean of a dataset is the sum of all values divided by the total number of values versus the
ii.	Mode is the value that appears the most often in a dataset, in contrast,
iii.	Median is the middle number in a sorted list of numbers of a dataset and can be more descriptive of that dataset than the average.